[rename] : test support 클래스들 파일 경로 수정
